Added Cmdlets for adding OCI server certificates to the trusted certification authorities
The cmdlet Import-OciServerCertificate was added to import the OCI
Server Certificate into the Trusted Root Certification Authority store
of the client. The Remove-OciServerCertificate was added to remove the
OCI Server Certificate from the Trusted Root Certification Authority
store. PowerShell shows a warning that there are security implications
when trusting a certificate which must be carefully considered!